put utils code in a new .c file
